rename Kafka stuff

+++ src/main/java/nl/craftsmen/orderservice/core/Order.java
@@ -7,6 +7,6 @@ public record Order(
         Long id,
         String customer,
         String name,
-        double price
+        long price
 ) {
 }
